Fixed a typo in the victory info. Tweaked the AI.
Fix a typo in CIV4VictoryInfo.xml which caused the space journey to be
twice as long as it is meant to be.

Adjusted the AI for full-culture in mastery victory, so that it abandons
culture4 when it is confident that it won't get any more legendary
cities.

(cherry picked from commit 2e2067913141e350161d312c657c93851d9e25bf)