Merge BULL pre-chop options
Originally by EmperorFool. Frequently (or, more so, urgently) requested by K-Mod players. Not safe for network games, so I'm letting the DLL treat pre-chop as disabled in network games. I've added a note about that to the hover text on the BUG menu.

The AdvCiv implementation of pre-chop, linked under this open K-Mod issue
https://github.com/karadoc/Civ4-K-Mod/issues/1
would be safe for multiplayer, but can't be used in conjunction with the Alt+C keyboard shortcut.