Fix crash when loading save from within a game
The crash occurs when the camera is near a city of a player with an id greater than the highest id of a living player in the savegame. I.e. only when the savegame has fewer players than the ongoing game.

Probably a Vanilla Civ 4 problem. Might be the one discussed here:
https://forums.civfanatics.com/threads/game-crash-when-loading-another-game-during-play.233021/