Enlarged replay screen
Should be easy to merge into other mods because most mods don't change the replay screen at all.

The resolution of the minimap can be increased through
MINIMAP_RENDER_SIZE in GlobalDefines
e.g. from 512 pixels (along the horizontal axis) to 1024.
But this setting also applies to the minimap on the main screen; I'm not sure if a higher resolution there won't (slightly) affect performance.