Staggered scoring for spaceship (Mastery)
At the request of Matty: 100 points for the first spaceship to arrive, 75 for second, 50 for third, 25 for fourth and beyond. See CvVictoryScreen.py for the Python integration. The point values are set in C++, but can be overwritten by defining integer GlobalDefines MASTERY_SPACE_SCORE_x in XML, where x is a number from 0 to 3, 0 corresponding to the first arrival and 3 to the fourth (and beyond).

Defeating a player whose spaceship has already arrived does not cause scores to be redistributed. When loading legacy savegames where one or multiple spaceships have already arrived, all are awarded 100 points (as before).